viewRegistraion
